requirements.txt
streamlit>=1.28.0
moviepy>=1.0.3
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0

